// src/app/models/DTO/BuyerAndBuyerCarDTO.ts
export class BuyerAndBuyerCarDTO {
    carSegment: String = '';
    mark: String = '';
    model: String = '';
    body: String = '';
    trunkCapacity: Number = null;
    drive: String = '';
    engine: Number = null;
    horsePower: Number = null;
    acceleration: Number = null;
    fuelConsumption: Number = null;
    price: Number = null;
    date: String = '';
    firstName: String = '';
    lastName: String = '';
}